requirements.txt
tensorflow>=2.8.0
pymatgen==2023.8.10
ase==3.21.1
scikit-learn>=1.2.1
numpy
h5py
pyyaml